Added main function, get matrix from file, started print matrix to file

+++ mat.h
@@ -6,11 +6,11 @@ int compare_matrices(double* a, double* b, int nRows, int nCols);
 double *read_matrix_from_file(const char *path);
 void print_matrix(double *a, int nrows, int ncols);
 
-int mmult(double *c, 
+double mmult(double *c,
           double *a, int aRows, int aCols, 
           double *b, int bRows, int bCols);
 
-int mmult_omp(double *c,
+double mmult_omp(double *c,
 		      const double *a, int aRows, int aCols,
 		      const double *b, int bRows, int bCols);
